Move processStyleTags to helpers
